Cast the packer's occupancy before inverting it
`_finalise` wrote the design PDB with `missing_atom_mask = 1.0 - mask`, where
`mask` comes straight from `_packed_mask` -- which deliberately returns the
packer's OWN tensor, dtype included. FaMPNN 0.3 returns that occupancy as
bool, and torch refuses `1.0 - <bool tensor>` outright:

    run_integrated_binder_matrix.py --targets PDL1 --lengths 80 --seeds 101
      --fampnn-checkpoint .../fampnn_0_3.pt --fampnn-variant 0.3

    File "scripts/run_integrated_binder_matrix.py", line 503, in _finalise
        "missing_atom_mask": (1.0 - mask).cpu(),
    RuntimeError: Subtraction, the `-` operator, with a bool tensor is not
    supported. If you are trying to invert a mask, use the `~` or
    `logical_not()` operator instead.

It fires after the event is recorded and every arm has been generated, on the
very last step of the cell, so the cost of hitting it is a whole trajectory.

Cast at the call site rather than inside `_packed_mask`: that helper's
contract is to hand back the packer's tensor unaltered, and both the fallback
path and `atom_mask_af2` rely on it. `.float()` is a no-op for a float
occupancy, so the previously intended behaviour is unchanged.

+++ scripts/run_integrated_binder_matrix.py
@@ -505,7 +505,13 @@ def _finalise(*, x0, products, structure, designer, adapters, args, out,
         "seq_mask": torch.ones(1, length),
         # The packer's OWN occupancy, not an all-zero mask: writing zeros
         # would claim every atom is present including ones it never built.
-        "missing_atom_mask": (1.0 - mask).cpu(),
+        # `.float()` because the occupancy comes straight from the packer and
+        # its dtype is the packer's choice: FaMPNN 0.3 returns it as bool,
+        # which `1.0 - x` refuses outright. Casting here rather than inside
+        # `_packed_mask` keeps that helper returning the packer's own tensor
+        # unaltered, which is what the fallback path and `atom_mask_af2`
+        # expect.
+        "missing_atom_mask": (1.0 - mask.float()).cpu(),
         "residue_index": topology.residue_index.reshape(1, -1).cpu().long(),
         "chain_index": topology.chain_index.reshape(1, -1).cpu().long(),
         "pred_aatype": products.aatype.cpu().long(),
